Greenfoot-source-2.1.0.zip

+++ bluej/src/bluej/terminal/Terminal.java
@@ -1,6 +1,6 @@
 /*
  This file is part of the BlueJ program. 
- Copyright (C) 1999-2010  Michael Kolling and John Rosenberg 
+ Copyright (C) 1999-2010,2011  Michael Kolling and John Rosenberg 
  
  This program is free software; you can redistribute it and/or 
  modify it under the terms of the GNU General Public License 
@@ -38,6 +38,7 @@ import java.awt.event.KeyListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.awt.print.PrinterJob;
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Reader;
@@ -63,6 +64,7 @@ import bluej.BlueJEventListener;
 import bluej.BlueJTheme;
 import bluej.Config;
 import bluej.debugger.Debugger;
+import bluej.debugger.DebuggerField;
 import bluej.debugger.DebuggerObject;
 import bluej.debugger.DebuggerTerminal;
 import bluej.debugmgr.ExecutionEvent;
@@ -267,10 +269,15 @@ public final class Terminal extends JFrame
     {
         initialise();
         String fileName = FileUtility.getFileName(this,
-                                 Config.getString("terminal.save.title"),
-                                 Config.getString("terminal.save.buttonText"),
-                                 null, false);
+                Config.getString("terminal.save.title"),
+                Config.getString("terminal.save.buttonText"),
+                null, false);
         if(fileName != null) {
+            File f = new File(fileName);
+            if (f.exists()){
+                if (DialogManager.askQuestion(this, "error-file-exists") != 0)
+                    return;
+            }
             try {
                 FileWriter writer = new FileWriter(fileName);
                 text.write(writer);
@@ -407,18 +414,20 @@ public final class Terminal extends JFrame
                 if (object != null) {
                     if (event.getClassName() != null && event.getMethodName() == null) {
                         // constructor call - the result object is the created object
-                        result = object.getGenClassName() + " result = ";
-                        result += "(new instance of " + object.getGenClassName() + ")";
+                        String genTypeStr = object.getGenType().toString();
+                        result = genTypeStr + " result = ";
+                        result += "(new instance of " + genTypeStr + ")";
                     }
                     else {
-                    	// if the method returns a void, we must handle it differently
+                        // if the method returns a void, we must handle it differently
                         if (object.isNullObject()) {
-                        	result = "void result";
+                            result = "void result";
                         }
                         else {
                             // other - the result object is a wrapper with a single result field
-                            result = object.getFieldValueTypeString(0) + " result = ";
-                            result += object.getFieldValueString(0);
+                            DebuggerField resultField = object.getField(0);
+                            result = resultField.getType() + " result = ";
+                            result += resultField.getValueString();
                         }
                     }
                 }
@@ -524,7 +533,7 @@ public final class Terminal extends JFrame
                     writeToTerminal("\n");
                     break;
 
-                case '\b':	// backspace
+                case '\b':  // backspace
                     if (buffer.backSpace()) {
                         try {
                             int length = text.getDocument().getLength();
@@ -535,8 +544,8 @@ public final class Terminal extends JFrame
                     }
                     break;
 
-                case '\r':	// carriage return
-                case '\n':	// newline
+                case '\r':  // carriage return
+                case '\n':  // newline
                     if (buffer.putChar('\n')) {
                         writeToTerminal(String.valueOf(ch));
                         buffer.notifyReaders();
@@ -552,7 +561,7 @@ public final class Terminal extends JFrame
             }
             break;
         }
-        event.consume();	// make sure the text area doesn't handle this
+        event.consume();    // make sure the text area doesn't handle this
     }
